+++ packages/carbon-graphs/src/js/helpers/localeFormat.js
@@ -1,0 +1,104 @@
+import * as d3 from 'd3';
+
+/**
+ * Creates format for locales like Spanish(es_ES), Portuguese(pt_BR) and Dutch(nl_NL).
+ *
+ * @private
+ * @param {object} date - date object which are going to assign for axis ticks.
+ * @returns {string} format to be displayed on axis.
+ */
+const getDefaultDateFormat = (date) => {
+  if (d3.timeSecond(date) < date) {
+    return d3.timeFormat('.%L')(date);
+  } if (d3.timeMinute(date) < date) {
+    return d3.timeFormat(':%S')(date);
+  } if (d3.timeHour(date) < date) {
+    return d3.timeFormat('%I:%M')(date);
+  } if (d3.timeDay(date) < date) {
+    return d3.timeFormat('%I %p')(date);
+  } if (d3.timeMonth(date) < date) {
+    return (d3.timeWeek(date) < date ? d3.timeFormat('%a %d')(date) : d3.timeFormat('%d %b')(date));
+  } if (d3.timeYear(date) < date) {
+    return d3.timeFormat('%B')(date);
+  }
+  return d3.timeFormat('%Y')(date);
+};
+
+/**
+ * Creates format for Germany(de_DE) locale.
+ *
+ * @private
+ * @param {object} date - date object which are going to assign for axis ticks.
+ * @returns {string} format to be displayed on axis.
+ */
+const getGermanyDateFormat = (date) => {
+  if (d3.timeSecond(date) < date) {
+    return d3.timeFormat('.%L')(date);
+  } if (d3.timeMinute(date) < date) {
+    return d3.timeFormat(':%S')(date);
+  } if (d3.timeHour(date) < date) {
+    return d3.timeFormat('%I:%M')(date);
+  } if (d3.timeDay(date) < date) {
+    return d3.timeFormat('%I %p')(date);
+  } if (d3.timeMonth(date) < date) {
+    return (d3.timeWeek(date) < date ? d3.timeFormat('%a %d.')(date) : d3.timeFormat('%d. %b')(date));
+  } if (d3.timeYear(date) < date) {
+    return d3.timeFormat('%B')(date);
+  }
+  return d3.timeFormat('%Y')(date);
+};
+
+/**
+ * Creates format for sweden(sv_SE) locale.
+ *
+ * @private
+ * @param {object} date - date object which are going to assign for axis ticks.
+ * @returns {string} format to be displayed on axis.
+ */
+const getSwedenDateFormat = (date) => {
+  if (d3.timeSecond(date) < date) {
+    return d3.timeFormat('.%L')(date);
+  } if (d3.timeMinute(date) < date) {
+    return d3.timeFormat(':%S')(date);
+  } if (d3.timeHour(date) < date) {
+    return d3.timeFormat('%I:%M')(date);
+  } if (d3.timeDay(date) < date) {
+    return d3.timeFormat('%I %p')(date);
+  } if (d3.timeMonth(date) < date) {
+    return (d3.timeWeek(date) < date ? d3.timeFormat('%a %_d')(date) : d3.timeFormat('%_d %b')(date));
+  } if (d3.timeYear(date) < date) {
+    return d3.timeFormat('%B')(date);
+  }
+  return d3.timeFormat('%Y')(date);
+};
+
+/**
+ * Creates format for French(fr_FR) locale.
+ *
+ * @private
+ * @param {object} date - date object which are going to assign for axis ticks.
+ * @returns {string} format to be displayed on axis.
+ */
+const getFrenchDateFormat = (date) => {
+  if (d3.timeSecond(date) < date) {
+    return d3.timeFormat('.%L')(date);
+  } if (d3.timeMinute(date) < date) {
+    return d3.timeFormat(':%S')(date);
+  } if (d3.timeHour(date) < date) {
+    return d3.timeFormat('%I:%M')(date);
+  } if (d3.timeDay(date) < date) {
+    return d3.timeFormat('%I %p')(date);
+  } if (d3.timeMonth(date) < date) {
+    return (d3.timeWeek(date) < date ? d3.timeFormat('%a %_d')(date) : d3.timeFormat('%d %b')(date));
+  } if (d3.timeYear(date) < date) {
+    return d3.timeFormat('%B')(date);
+  }
+  return d3.timeFormat('%Y')(date);
+};
+
+export {
+  getDefaultDateFormat,
+  getGermanyDateFormat,
+  getSwedenDateFormat,
+  getFrenchDateFormat,
+};