+++ packages/carbon-graphs/src/js/locale/de-DE.js
@@ -16,7 +16,7 @@ export default {
     'Freitag',
     'Samstag',
   ],
-  shortDays: ['So', 'Mo', 'Di', 'Mi', 'Do', 'Fr', 'Sa'],
+  shortDays: ['So,', 'Mo,', 'Di,', 'Mi,', 'Do,', 'Fr,', 'Sa,'],
   months: [
     'Januar',
     'Februar',